added start of assembly code partition/persistent/package functionality
have been wondering it would be convenient if we would have a way to have a way to be able to organize assembly code into queriable units. For a number of reasons: So we can compile different projects next to each other, so we can test parts of the application more easily, to make incremental compilation possible... This is a first stab/stage. code can be partitioned in different packages, and then in different blocks within those packages, to accommodate for different memory addresses. Needs to be wired into the assembler better, but it works quite nicely (better than expected) for my debugging purposes.

darcs-hash:20080602040306-f86be-86e81f6b2fed86fd21ce386d1d3cf47a00efa5e2